chore : 공백 제거 및 메서드 반환형 변환

+++ src/main/java/lotto/domain/PrizeType.java
@@ -23,7 +23,7 @@ public enum PrizeType {
         return prizeMoney;
     }
 
-    public double getMatchCount(){
+    public int getMatchCount(){
         return matchCount;
     }
